backend/requirements/development.txt
# Include production dependencies
-r production.txt

# Development Tools
ipython==8.20.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Redis for Channels (for WebSocket communication across processes)
channels-redis==4.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3

# Documentation
sphinx==7.2.6
